201 OneToMany Polymorphic associating #144

+++ database/migrations/2022_09_05_015902_add_polymorph_to_comments_table.php
@@ -29,7 +29,7 @@ class AddPolymorphToCommentsTable extends Migration
     {
         Schema::table('comments', function (Blueprint $table) {
             $table->dropMorphs('commentable');
-            $table->unsignedInteger('blog_post_id')->index();
+            $table->unsignedBigInteger('blog_post_id')->index();
             $table->foreign('blog_post_id')->references('id')->on('blog_posts');
         });
     }
